backend/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.4.0
idna==3.10
psycopg2==2.9.10
python-dotenv==1.1.0
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
gunicorn==23.0.0
whitenoise==6.8.2
